+++ dev.R
@@ -16,3 +16,4 @@ dings$char_get_code()
 # plot text - polygons?
 
 # extensive testing on init!!!
+